src/backend/requirements.txt
pandas
requests
scikit-learn
fastapi[all]
sqlalchemy
uvicorn
langchain
langgraph
langchain-google-genai
# langchain-openai
langchain-chroma
pydantic
pymysql
rq
redis
argon2-cffi
PyJWT
cryptography
jira2markdown
python-dotenv


# nltk
# sentence-transformers
# torch